requirements.txt
# ═══════════════════════════════════════════════════════════════════════════
# GingivaGen 2.0 — Multi-Material Hybrid Scaffold Pipeline
# ═══════════════════════════════════════════════════════════════════════════

# Core scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Mesh I/O & manipulation
trimesh>=3.21.0
meshlib>=3.1.0

# TPMS lattice generation (LisbonTPMS)
# Install from: https://github.com/SoftwareImpacts/SIMPAC-2025-26
# pip install git+https://github.com/SoftwareImpacts/SIMPAC-2025-26.git

# Pore & porosity analysis
porespy>=2.0.0
scikit-image>=0.20.0

# Multi-material G-code generation
fullcontrol>=0.4.0

# Neural segmentation (3DTeethSAM) dependencies
torch>=2.0.0
torchvision>=0.15.0
pytorch3d>=0.7.0

# Visualization (optional but recommended)
pyvista>=0.38.0
matplotlib>=3.7.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Testing
pytest>=7.0.0
